add home, header and footer
