+++ schedulerApp/frontend/package.json
@@ -13,6 +13,7 @@
     "react-bootstrap": "^2.8.0",
     "react-bootstrap-validation": "^0.1.11",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
     "reactstrap": "^8.9.0",
     "web-vitals": "^2.1.4"
