+++ schedulerApp/frontend/src/Pages/Home/Banner/Banner.css
@@ -1,0 +1,17 @@
+/* Banner */
+
+.banner {
+    width: 100%;
+    height: 200px;
+    background-color: #004caa;
+    color: #0e0202;
+    text-align: center;
+    padding-top: 50px;
+  }
+  
+  .banner h1 {
+    font-size: 60px;
+    font-weight: bold;
+    margin-bottom: 0;
+  }
+
