+++ schedulerApp/frontend/src/Pages/Home/Banner/Banner.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import './Banner.css';
+
+const Banner = () => {
+  return (
+    <div class = "banner">
+        <h1>Projeto MC426 </h1>
+    </div>
+  );
+}
+
+export default Banner;
