+++ schedulerApp/frontend/src/Pages/Home/Home/Home.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import Banner from '../Banner/Banner';
+import HomeInfo from '../HomeInfo/HomeInfo'
+
+const Home = () => {
+  return (
+    <>
+       <Banner />
+       <HomeInfo />
+    </>
+  );
+
+}
+
+export default Home;
