+++ schedulerApp/frontend/src/Pages/Home/HomeInfo/HomeInfo.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import './HomeInfo.css';
+
+const HomeInfo = () => {
+  return (
+    <div class>
+        <h1>Projeto MC426 </h1>
+        <h2> Teste teste</h2>
+    </div>
+  );
+}
+
+export default HomeInfo;
