+++ schedulerApp/frontend/src/routes.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { Route, BrowserRouter } from "react-router-dom";
+
+import Home from "./Pages/Home/Home";
+
+const Routes = () => {
+   return(
+       <BrowserRouter>
+           <Route component = { Home }  path="/" exact />
+       </BrowserRouter>
+   );
+}
+
+export default Routes;
